Simple Web Server
Flask allows for Python to be run in place of PHP. I went through the leg work of getting this small reproducible example running for MAC. So long as you have pip3 or pip and IPython notebook or Jupyter notebook you should be fine to use this. This will let us run our sentiment analysis on our project (or another python predictive model on it) behind the scenes when we start predicting chapters.